+++ server/oneyearfourcut/src/test/java/com/codestates/mainproject/oneyearfourcut/domain/comment/controller/CommentControllerTest.java
@@ -1,660 +1,559 @@
-//package com.codestates.mainproject.oneyearfourcut.domain.comment.controller;
-//
-//import com.codestates.mainproject.oneyearfourcut.domain.artwork.entity.Artwork;
-//import com.codestates.mainproject.oneyearfourcut.domain.artwork.entity.ArtworkStatus;
-//import com.codestates.mainproject.oneyearfourcut.domain.artwork.repository.ArtworkRepository;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentArtworkResDto;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentGalleryResDto;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentRequestDto;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.Comment;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.CommentStatus;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.repository.CommentRepository;
-//import com.codestates.mainproject.oneyearfourcut.domain.comment.service.CommentService;
-//import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.Gallery;
-//import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.GalleryStatus;
-//import com.codestates.mainproject.oneyearfourcut.domain.gallery.repository.GalleryRepository;
-//import com.codestates.mainproject.oneyearfourcut.domain.member.entity.Member;
-//import com.codestates.mainproject.oneyearfourcut.domain.member.entity.Role;
-//import com.codestates.mainproject.oneyearfourcut.domain.member.repository.MemberRepository;
-//import com.codestates.mainproject.oneyearfourcut.global.config.auth.jwt.JwtTokenizer;
-//import com.codestates.mainproject.oneyearfourcut.global.page.*;
-//import com.google.gson.Gson;
-//import org.junit.jupiter.api.Test;
-//import org.mockito.Mockito;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-//import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-//import org.springframework.boot.test.context.SpringBootTest;
-//import org.springframework.boot.test.mock.mockito.MockBean;
-//import org.springframework.data.domain.Page;
-//import org.springframework.data.domain.PageRequest;
-//import org.springframework.http.MediaType;
-//import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
-//import org.springframework.restdocs.payload.JsonFieldType;
-//import org.springframework.security.test.context.support.WithMockUser;
-//import org.springframework.test.web.servlet.MockMvc;
-//import org.springframework.test.web.servlet.ResultActions;
-//import org.springframework.transaction.annotation.Transactional;
-//
-//import java.time.LocalDateTime;
-//import java.util.List;
-//import java.util.Optional;
-//import java.util.stream.Collectors;
-//
-//import static com.codestates.mainproject.oneyearfourcut.domain.comment.entity.CommentStatus.VALID;
-//import static com.codestates.mainproject.oneyearfourcut.domain.member.entity.MemberStatus.ACTIVE;
-//import static com.codestates.mainproject.oneyearfourcut.global.util.ApiDocumentUtils.getRequestPreProcessor;
-//import static com.codestates.mainproject.oneyearfourcut.global.util.ApiDocumentUtils.getResponsePreProcessor;
-//import static org.mockito.ArgumentMatchers.*;
-//import static org.mockito.BDDMockito.given;
-//import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
-//import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
-//import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-//import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-//import static org.springframework.restdocs.request.RequestDocumentation.*;
-//import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-//import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-//
-//@Transactional
-//@SpringBootTest
-//@AutoConfigureMockMvc
-//@AutoConfigureRestDocs
-//@WithMockUser(username = "test@gmail.com", password = "0000")
-//class CommentControllerTest {
-//
-//    @Autowired
-//    private MockMvc mockMvc;
-//    @Autowired
-//    private JwtTokenizer jwtTokenizer;
-//    @Autowired
-//    private MemberRepository memberRepository;
-//    @Autowired
-//    private GalleryRepository galleryRepository;
-//    @Autowired
-//    private CommentRepository commentRepository;
-//    private ArtworkRepository artworkRepository;
-//    @MockBean
-//    private CommentService commentService;
-//
-//    @Autowired
-//    private Gson gson;
-//
-//    @Test
-//    void  testPostCommentOnGallery() throws Exception{
-//        //given
-//        Member member = memberRepository.save(Member.builder()
-//                .nickname("test1")
-//                .email("kang@gmail.com")
-//                .role(Role.USER)
-//                .profile("/path")
-//                .status(ACTIVE)
-//                .build());
-//        String jwt = jwtTokenizer.testJwtGenerator(member);
-//        Gallery gallery = galleryRepository.save(new Gallery(1L));
-//
-//        CommentRequestDto requestDto = CommentRequestDto.builder()
-//                .content("댓글입니다")
-//                .build();
-//        String gsonContent = gson.toJson(requestDto);
-//
-//        CommentGalleryResDto responseDto = CommentGalleryResDto.builder()
-//                .createdAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-//                .modifiedAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-//                .commentId(1L)
-//                .memberId(member.getMemberId())
-//                .nickname(member.getNickname())
-//                .content(requestDto.getContent())
-//                .artworkId(null)
-//                .build();
-//
-//        given(this.commentService.createCommentOnGallery(Mockito.any(requestDto.getClass()),
-//                Mockito.any( member.getMemberId().getClass() ),
-//                Mockito.any( gallery.getGalleryId().getClass() ) ) )
-//                .willReturn(new CommentGalleryHeadDto<>(1L,responseDto));
-//
-//        //when
-//        ResultActions actions = mockMvc.perform(
-//                RestDocumentationRequestBuilders.post("/galleries/{galleryId}/comments", gallery.getGalleryId())
-//                        .header("Authorization", jwt)
-//                        .accept(MediaType.APPLICATION_JSON)
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                        .content(gsonContent)
-//                        /*.with(csrf())*/
-//        );
-//        commentService.createCommentOnGallery(requestDto, member.getMemberId(), gallery.getGalleryId());
-//
-//        //then
-//        actions.andExpect(status().isCreated())
-//                .andExpect(jsonPath("$.galleryId").value( gallery.getGalleryId()))
-//                .andExpect(jsonPath("$.commentList.createdAt").value("2022-11-25T11:09:24.94"))
-//                .andExpect(jsonPath("$.commentList.modifiedAt").value("2022-11-25T11:09:24.94"))
-//                .andExpect(jsonPath("$.commentList.commentId").value(1L))
-//                .andExpect(jsonPath("$.commentList.memberId").value(member.getMemberId()))
-//                .andExpect(jsonPath("$.commentList.nickname").value("test1"))
-//                .andExpect(jsonPath("$.commentList.content").value("댓글입니다"))
-//                .andExpect(jsonPath("$.commentList.artworkId").doesNotExist())
-//
-//                .andDo(document(
-//                        "postCommentOnGallery",
-//                        getRequestPreProcessor(),
-//                        getResponsePreProcessor(),
-//                        requestHeaders(
-//                                List.of(
-//                                        headerWithName("Authorization").description("JWT - Access Token")
-//                                )
-//                        ),
-//                                pathParameters(
-//                                        parameterWithName("galleryId").description("갤러리 Id")),
-//                        requestFields(
-//                                List.of(
-//                                        fieldWithPath("content").type(JsonFieldType.STRING).description("댓글 내용")
-//                                )
-//                        )
-//                        , responseFields(
-//                                List.of(
-//                                        fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
-//                                        fieldWithPath("commentList.createdAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                        fieldWithPath("commentList.modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                        fieldWithPath("commentList.commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
-//                                        fieldWithPath("commentList.memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
-//                                        fieldWithPath("commentList.nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
-//                                        fieldWithPath("commentList.content").type(JsonFieldType.STRING).description("댓글 내용"),
-//                                        fieldWithPath("commentList.artworkId").type(JsonFieldType.NUMBER).description("작품 ID").optional(),
-//                                        fieldWithPath("commentList.imagePath").type(JsonFieldType.NULL).description("작품 이미지 url")
-//                                )
-//                        )
-//                )
-//                );
-//    }
-//    @Test
-//    void testPostCommentOnArtwork() throws Exception {
-//        //given
-//        Member member = memberRepository.save(Member.builder()
-//                .nickname("test1")
-//                .email("kang@gmail.com")
-//                .role(Role.USER)
-//                .profile("/path")
-//                .status(ACTIVE)
-//                .build());
-//        String jwt = jwtTokenizer.testJwtGenerator(member);
-//        Gallery gallery = galleryRepository.save(new Gallery(1L));
-//
-//        CommentRequestDto requestDto = CommentRequestDto.builder()
-//                .content("댓글입니다")
-//                .build();
-//        String gsonContent = gson.toJson(requestDto);
-//
-//        CommentArtworkResDto responseDto = CommentArtworkResDto.builder()
-//                .createdAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-//                .modifiedAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-//                .commentId(1L)
-//                .memberId(member.getMemberId())
-//                .nickname(member.getNickname())
-//                .content(requestDto.getContent())
-//                .build();
-//
-//        Long artworkId = 1L;
-//
-//        given(this.commentService.createCommentOnArtwork(Mockito.any(requestDto.getClass()),
-//                Mockito.any( member.getMemberId().getClass() ),
-//                Mockito.any(artworkId.getClass()),
-//                Mockito.any( gallery.getGalleryId().getClass() ) ) )
-//                .willReturn(new CommentArtworkHeadDto<>(1L,1L,responseDto));
-//
-//        //when
-//        ResultActions actions = mockMvc.perform(
-//                RestDocumentationRequestBuilders
-//                        .post("/galleries/{galleryId}/artworks/{artworkId}/comments",
-//                                gallery.getGalleryId(), 1L)
-//                        .header("Authorization", jwt)
-//                        .accept(MediaType.APPLICATION_JSON)
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                        .content(gsonContent)
-//                /*.with(csrf())*/
-//        );
-//        commentService.createCommentOnArtwork(requestDto, member.getMemberId(), 1L, gallery.getGalleryId());
-//
-//        //then
-//        actions.andExpect(status().isCreated())
-//                .andExpect(jsonPath("$.galleryId").value( gallery.getGalleryId()))
-//                .andExpect(jsonPath("$.artworkId").value( 1L ))
-//                .andExpect(jsonPath("$.commentList.createdAt").value("2022-11-25T11:09:24.94"))
-//                .andExpect(jsonPath("$.commentList.modifiedAt").value("2022-11-25T11:09:24.94"))
-//                .andExpect(jsonPath("$.commentList.commentId").value(1L))
-//                .andExpect(jsonPath("$.commentList.memberId").value(member.getMemberId()))
-//                .andExpect(jsonPath("$.commentList.nickname").value("test1"))
-//                .andExpect(jsonPath("$.commentList.content").value("댓글입니다"))
-//
-//
-//                .andDo(document(
-//                        "postCommentOnArtwork",
-//                        getRequestPreProcessor(),
-//                        getResponsePreProcessor(),
-//                        requestHeaders(
-//                                List.of(
-//                                        headerWithName("Authorization").description("JWT - Access Token")
-//                                )
-//                        )
-//                                ,pathParameters(
-//                                        parameterWithName("galleryId").description("갤러리 Id"),
-//                                        parameterWithName("artworkId").description("작품 Id")),
-//
-//                        requestFields(
-//                                List.of(
-//                                        fieldWithPath("content").type(JsonFieldType.STRING).description("댓글 내용")
-//                                )
-//                        )
-//                        , responseFields(
-//                                List.of(
-//                                        fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
-//                                        fieldWithPath("artworkId").type(JsonFieldType.NUMBER).description("작품 ID"),
-//                                        fieldWithPath("commentList.createdAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                        fieldWithPath("commentList.modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                        fieldWithPath("commentList.commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
-//                                        fieldWithPath("commentList.memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
-//                                        fieldWithPath("commentList.nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
-//                                        fieldWithPath("commentList.content").type(JsonFieldType.STRING).description("댓글 내용")
-//
-//                                )
-//                        )
-//                )
-//                );
-//
-//    }
-//    @Test
-//    void testGetGalleryComment() throws Exception{
-//        //given
-//        Member member = memberRepository.save(Member.builder()
-//                .nickname("test1")
-//                .email("kang@gmail.com")
-//                .role(Role.USER)
-//                .profile("/path")
-//                .status(ACTIVE)
-//                .build());
-//
-//        Gallery gallery = galleryRepository.save(Gallery.builder()
-//                .title("나의 전시관")
-//                .content("안녕하세요")
-//                .member(member)
-//                .status(GalleryStatus.OPEN)
-//                .build());
-//
-//        int page = 1; int size = 10;
-//        PageRequest pr = PageRequest.of(page - 1, size);
-//        Page<Comment> commentPage = commentRepository.findAllByCommentStatusAndGallery_GalleryIdAndArtwork_StatusOrderByCommentIdDesc(VALID,1L, ArtworkStatus.REGISTRATION, pr);
-//
-////        given(commentRepository
-////                .findAllByCommentStatusAndGallery_GalleryIdAndArtwork_StatusOrderByCommentIdDesc(eq(VALID), eq(1L), eq(ArtworkStatus.REGISTRATION), eq(pr)))
-////                .willReturn(commentPage);
-//
-////        given(this.commentService.findCommentByPage(
-////                Mockito.any( gallery.getGalleryId().getClass() ),
-////                Mockito.any(),
-////                eq(page),
-////                eq(size)))
-////                .willReturn(commentPage);
-////
-////        Page<Comment> commentPage1 = commentService.findCommentByPage(1L, null, page, size);
-////        List<Comment> commentList = commentPage1.getContent();
-//        List<Comment> commentList = commentPage.getContent();
-//
-//        //임시로 생성, 추후 테스트 코드 고쳐야함
-//        List<CommentGalleryResDto> collect = commentList.stream()
-//                .map(comment -> {
-//                    return comment.toCommentGalleryResponseDto("/test.jpg"); //
-//                })
-//                .collect(Collectors.toList());
-//        //
-//        PageInfo<Object> pageInfo = new PageInfo<>(page, size, (int) commentPage.getTotalElements(), commentPage.getTotalPages());
-//
-//        given(this.commentService.getGalleryCommentPage(
-//                Mockito.any( gallery.getGalleryId().getClass() ),
-//                eq(page),
-//                eq(size)))
-//                .willReturn(new CommentGalleryPageResponseDto<>(gallery.getGalleryId(),collect,pageInfo));
-//
-//        //when
-//        ResultActions actions = mockMvc.perform(
-//                RestDocumentationRequestBuilders
-//                        .get("/galleries/{galleryId}/comments",
-//                               gallery.getGalleryId())
-//                        .param("page", String.valueOf(1))
-//                        .accept(MediaType.APPLICATION_JSON)
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                /*.with(csrf())*/
-//        );
-//        commentService.getGalleryCommentPage(gallery.getGalleryId(),1,10);
-//
-//        //then
-//        actions.andExpect(status().isOk())
-//                .andExpect(jsonPath("$.galleryId").value( gallery.getGalleryId()))
-//                .andExpect(jsonPath("$.commentList[0].createdAt").value("2022-11-12T23:41:58.764644"))
-//                .andExpect(jsonPath("$.commentList[0].modifiedAt").value("2022-11-16T23:41:57.764644"))
-//                .andExpect(jsonPath("$.commentList[0].commentId").value(2L))
-//                .andExpect(jsonPath("$.commentList[0].memberId").value(1L))
-//                .andExpect(jsonPath("$.commentList[0].nickname").value("gallery"))
-//                .andExpect(jsonPath("$.commentList[0].content").value("comment2"))
-//                .andExpect(jsonPath("$.commentList[0].artworkId").doesNotExist())
-//
-//                .andExpect(jsonPath("$.commentList[1].createdAt").value("2022-11-12T23:41:58.764644"))
-//                .andExpect(jsonPath("$.commentList[1].modifiedAt").value("2022-11-16T23:41:57.764644"))
-//                .andExpect(jsonPath("$.commentList[1].commentId").value(2L))
-//                .andExpect(jsonPath("$.commentList[1].memberId").value(1L))
-//                .andExpect(jsonPath("$.commentList[1].nickname").value("gallery"))
-//                .andExpect(jsonPath("$.commentList[1].content").value("comment2"))
-//                .andExpect(jsonPath("$.commentList[1].artworkId").value(2L))
-//
-//                .andExpect(jsonPath("$.pageInfo.page").value("1"))
-//                .andExpect(jsonPath("$.pageInfo.size").value("10"))
-//                .andExpect(jsonPath("$.pageInfo.totalElements").value("3"))
-//                .andExpect(jsonPath("$.pageInfo.totalPages").value("1"))
-//
-//
-//                .andDo(document(
-//                                "getCommentOnGallery",
-//                                getRequestPreProcessor(),
-//                                getResponsePreProcessor(),
-//                                requestParameters(parameterWithName("page").description("페이지 수")
-//                        )
-//                        ,pathParameters(
-//                                parameterWithName("galleryId").description("갤러리 Id"))
-//                                , responseFields(
-//                                        List.of(
-//                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
-//                                                fieldWithPath("commentList[].createdAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                                fieldWithPath("commentList[].modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                                fieldWithPath("commentList[].commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
-//                                                fieldWithPath("commentList[].memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
-//                                                fieldWithPath("commentList[].nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
-//                                                fieldWithPath("commentList[].content").type(JsonFieldType.STRING).description("댓글 내용"),
-//                                                fieldWithPath("commentList[].artworkId").type(JsonFieldType.NUMBER).description("작품 ID").optional(),
-//                                                fieldWithPath("commentList[].imagePath").type(JsonFieldType.STRING).description("작품 이미지 URL (없으면 null)"),
-//
-//                                                fieldWithPath("pageInfo.page").type(JsonFieldType.NUMBER).description("현 페이지 위치"),
-//                                                fieldWithPath("pageInfo.size").type(JsonFieldType.NUMBER).description("페이지 내용 크기"),
-//                                                fieldWithPath("pageInfo.totalElements").type(JsonFieldType.NUMBER).description("총 댓글 수"),
-//                                                fieldWithPath("pageInfo.totalPages").type(JsonFieldType.NUMBER).description("총 페이지 수")
-//                                        )
-//                                )
-//                        )
-//                );
-//    }
-//    @Test
-//    void testGetArtworkComment() throws Exception {
-//        //given
-//        Member member = memberRepository.save(Member.builder()
-//                .nickname("test1")
-//                .email("kang@gmail.com")
-//                .role(Role.USER)
-//                .profile("/path")
-//                .status(ACTIVE)
-//                .build());
-//
-//        Gallery gallery = galleryRepository.save(Gallery.builder()
-//                .title("나의 전시관")
-//                .content("안녕하세요")
-//                .member(member)
-//                .status(GalleryStatus.OPEN)
-//                .build());
-//
-//        int page = 1; int size = 10; Long artworkId = 1L;
-//        PageRequest pr = PageRequest.of(page - 1, size);
-//        Page<Comment> commentPage = commentRepository.findAllByCommentStatusAndArtwork_ArtworkIdOrderByCommentIdDesc(VALID,1L, pr);
-//
-////        given(this.commentRepository.findAllByCommentStatusAndArtwork_ArtworkIdOrderByCommentIdDesc())
-////        given(this.commentService.findCommentByPage(
-////                Mockito.any( gallery.getGalleryId().getClass() ),
-////                Mockito.any(),
-////                eq(page),
-////                eq(size)))
-////                .willReturn(commentPage);
-////
-////        Page<Comment> commentPage1 = commentService.findCommentByPage(1L, 1L, page, size);
-////        List<Comment> commentList = commentPage1.getContent();
-////        PageInfo<Object> pageInfo = new PageInfo<>(page, size, (int) commentPage1.getTotalElements(), commentPage1.getTotalPages());
-////        List<CommentArtworkResDto> response = CommentArtworkResDto.toCommentArtworkResponseDtoList(commentList);
-////
-////        given(this.commentService.getArtworkCommentPage(
-////                Mockito.any( gallery.getGalleryId().getClass() ), eq(artworkId),
-////                eq(page),
-////                eq(size)))
-////                .willReturn(new CommentArtworkPageResponseDto<>(gallery.getGalleryId(), 1L ,response,pageInfo));
-//
-//        //when
-//        ResultActions actions = mockMvc.perform(
-//                RestDocumentationRequestBuilders
-//                        .get("/galleries/{galleryId}/artworks/{artworkId}/comments",
-//                                gallery.getGalleryId(), 1L)
-//                        .param("page", String.valueOf(1))
-//                        .accept(MediaType.APPLICATION_JSON)
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                /*.with(csrf())*/
-//        );
-//        commentService.getArtworkCommentPage(gallery.getGalleryId(), 1L,1,10);
-//
-//        //then
-//        actions.andExpect(status().isOk())
-//                .andExpect(jsonPath("$.galleryId").value( gallery.getGalleryId()))
-//                .andExpect(jsonPath("$.artworkId").value( 1L ))
-//                .andExpect(jsonPath("$.commentList[0].commentId").value(1L))
-//                .andExpect(jsonPath("$.commentList[0].memberId").value(1L))
-//                .andExpect(jsonPath("$.commentList[0].nickname").value("gallery"))
-//                .andExpect(jsonPath("$.commentList[0].content").value("comment1댓글대스"))
-//
-//                .andExpect(jsonPath("$.pageInfo.page").value("1"))
-//                .andExpect(jsonPath("$.pageInfo.size").value("10"))
-//                .andExpect(jsonPath("$.pageInfo.totalElements").value("1"))
-//                .andExpect(jsonPath("$.pageInfo.totalPages").value("1"))
-//
-//
-//                .andDo(document(
-//                                "getCommentOnArtwork",
-//                                getRequestPreProcessor(),
-//                                getResponsePreProcessor(),
-//                                requestParameters(parameterWithName("page").description("페이지 수")
-//                                )
-//                                ,pathParameters(
-//                                        parameterWithName("galleryId").description("갤러리 Id"),
-//                                parameterWithName("artworkId").description("작품 Id"))
-//                                , responseFields(
-//                                        List.of(
-//                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
-//                                                fieldWithPath("artworkId").type(JsonFieldType.NUMBER).description("작품 ID"),
-//                                                fieldWithPath("commentList[].createdAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                                fieldWithPath("commentList[].modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                                fieldWithPath("commentList[].commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
-//                                                fieldWithPath("commentList[].memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
-//                                                fieldWithPath("commentList[].nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
-//                                                fieldWithPath("commentList[].content").type(JsonFieldType.STRING).description("댓글 내용"),
-//
-//                                                fieldWithPath("pageInfo.page").type(JsonFieldType.NUMBER).description("현 페이지 위치"),
-//                                                fieldWithPath("pageInfo.size").type(JsonFieldType.NUMBER).description("페이지 내용 크기"),
-//                                                fieldWithPath("pageInfo.totalElements").type(JsonFieldType.NUMBER).description("총 댓글 수"),
-//                                                fieldWithPath("pageInfo.totalPages").type(JsonFieldType.NUMBER).description("총 페이지 수")
-//                                        )
-//                                )
-//                        )
-//                );
-//    }
-//    @Test
-//    void testPatchComment() throws Exception {
-//        //given
-//        Member member3 = memberRepository.save(Member.builder()
-//                .nickname("test1")
-//                .email("kang@gmail.com")
-//                .role(Role.USER)
-//                .profile("/path")
-//                .status(ACTIVE)
-//                .build());
-//        String jwt = jwtTokenizer.testJwtGenerator(member3);
-//
-//        Gallery gallery3 = galleryRepository.save(Gallery.builder()
-//                .title("나의 전시관")
-//                .content("안녕하세요")
-//                .member(member3)
-//                .status(GalleryStatus.OPEN)
-//                .build());
-//        Artwork artwork3 = Artwork.builder()
-//                .title("나의 작품")
-//                .content("설명글")
-//                .build();
-//        artwork3.setMember(member3);
-//        artwork3.setGallery(gallery3);
-//        artwork3.setImagePath("/test.jpg");
-//
-//        Comment comment = commentRepository.save(Comment.builder()
-//                .content("오리지널댓글")
-//                .member(member3)
-//                .gallery(gallery3)
-//                .artwork(artwork3)
-//                .build());
-//        System.out.println("comment.getCommentId() = " + comment.getCommentId());
-//        CommentRequestDto requestDto = CommentRequestDto.builder()
-//                .content("수정댓글입니다.")
-//                .build();
-//        String gsonContent = gson.toJson(requestDto);
-//
-//        Long commentId = 6L;
-//
-//        given(this.commentService.findComment(commentId)).willReturn(comment);
-//
-//        Optional<Comment> foundComment = commentRepository.findById(commentId);
-//
-//        CommentGalleryResDto responseDto = CommentGalleryResDto.builder()
-//                .createdAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-//                .modifiedAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-//                .commentId(6L)
-//                .memberId(member3.getMemberId())
-//                .nickname(member3.getNickname())
-//                .content(requestDto.getContent())
-//                .artworkId(null)
-//                .build();
-//
-//        given(this.commentService.modifyComment(Mockito.any( gallery3.getGalleryId().getClass()),
-//                eq(commentId),
-//                Mockito.any( requestDto.getClass() ),
-//                Mockito.any( member3.getMemberId().getClass() )))
-//                .willReturn(new CommentGalleryHeadDto<>( gallery3.getGalleryId() , responseDto));
-//
-//        //when
-//        ResultActions actions = mockMvc.perform(
-//                RestDocumentationRequestBuilders
-//                        .patch("/galleries/{galleryId}/comments/{commentId}",
-//                                gallery3.getGalleryId(), comment.getCommentId())
-//                        .header("Authorization", jwt)
-//                        .accept(MediaType.APPLICATION_JSON)
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                        .content(gsonContent)
-//                /*.with(csrf())*/
-//        );
-//        commentService.modifyComment(gallery3.getGalleryId(),responseDto.getCommentId(), requestDto, member3.getMemberId());
-//        System.out.println("gallery.getGalleryId() = " + gallery3.getGalleryId());
-//
-//        //then
-//        actions.andExpect(status().isOk())
-//                .andExpect(jsonPath("$.galleryId").value( gallery3.getGalleryId()))
-//                .andExpect(jsonPath("$.commentList.commentId").value( 6L ))
-//                .andExpect(jsonPath("$.commentList.memberId").value(member3.getMemberId()))
-//                .andExpect(jsonPath("$.commentList.nickname").value("test1"))
-//                .andExpect(jsonPath("$.commentList.content").value("수정댓글입니다."))
-//                .andExpect(jsonPath("$.commentList.artworkId").doesNotExist())
-//
-//                .andDo(document(
-//                                "patchComment",
-//                                getRequestPreProcessor(),
-//                                getResponsePreProcessor(),
-//                                requestHeaders(
-//                                        List.of(
-//                                                headerWithName("Authorization").description("JWT - Access Token")
-//                                        )
-//                                ),
-//                                pathParameters(
-//                                        parameterWithName("galleryId").description("갤러리 Id"),
-//                                        parameterWithName("commentId").description("댓글 Id")),
-//                                requestFields(
-//                                        List.of(
-//                                                fieldWithPath("content").type(JsonFieldType.STRING).description("댓글 내용")
-//                                        )
-//                                )
-//                                , responseFields(
-//                                        List.of(
-//                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
-//                                                fieldWithPath("commentList.createdAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                                fieldWithPath("commentList.modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
-//                                                fieldWithPath("commentList.commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
-//                                                fieldWithPath("commentList.memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
-//                                                fieldWithPath("commentList.nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
-//                                                fieldWithPath("commentList.content").type(JsonFieldType.STRING).description("댓글 내용"),
-//                                                fieldWithPath("commentList.artworkId").type(JsonFieldType.NUMBER).description("작품 ID").optional(),
-//                                                fieldWithPath("commentList.imagePath").type(JsonFieldType.NULL).description("작품 이미지 url")
-//                                        )
-//                                )
-//                        )
-//                );
-//    }
-//    @Test
-//    void testDeleteComment() throws Exception{
-//        //given
-//        Member member = memberRepository.save(Member.builder()
-//                .nickname("test1")
-//                .email("kang@gmail.com")
-//                .role(Role.USER)
-//                .profile("/path")
-//                .status(ACTIVE)
-//                .build());
-//        String jwt = jwtTokenizer.testJwtGenerator(member);
-//
-//        Gallery gallery = galleryRepository.save(Gallery.builder()
-//                .title("나의 전시관")
-//                .content("안녕하세요")
-//                .member(member)
-//                .status(GalleryStatus.OPEN)
-//                .build());
-//        Artwork artwork = Artwork.builder()
-//                .title("나의 작품")
-//                .content("설명글")
-//                .build();
-//        artwork.setMember(member);
-//        artwork.setGallery(gallery);
-//        artwork.setImagePath("/test.jpg");
-//
-//        Comment comment = commentRepository.save(Comment.builder()
-//                .commentId(6L)
-//                .content("댓글")
-//                .member(member)
-//                .gallery(gallery)
-//                .artwork(artwork)
-//                .build());
-//
-//        //when
-//        ResultActions actions = mockMvc.perform(
-//                RestDocumentationRequestBuilders
-//                        .delete("/galleries/{galleryId}/comments/{commentId}",
-//                                gallery.getGalleryId(), comment.getCommentId())
-//                        .header("Authorization", jwt)
-//                        .accept(MediaType.APPLICATION_JSON)
-//                        .contentType(MediaType.APPLICATION_JSON)
-//                /*.with(csrf())*/
-//        );
-//
-//        //then
-//        actions.andExpect(status().isNoContent())
-//                .andDo(document(
-//                        "deleteComment",
-//                        getRequestPreProcessor(),
-//                        getResponsePreProcessor(),
-//                        requestHeaders(
-//                                List.of(
-//                                        headerWithName("Authorization").description("JWT - Access Token")
-//                                )
-//                        ),
-//                        pathParameters(
-//                                parameterWithName("galleryId").description("갤러리 Id"),
-//                                parameterWithName("commentId").description("댓글 Id"))
-//                        )
-//
-//                );
-//
-//    }
-//
-//}
+package com.codestates.mainproject.oneyearfourcut.domain.comment.controller;
+
+import com.codestates.mainproject.oneyearfourcut.domain.artwork.entity.Artwork;
+import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentArtworkResDto;
+import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentGalleryResDto;
+import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentRequestDto;
+import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.Comment;
+import com.codestates.mainproject.oneyearfourcut.domain.comment.service.CommentService;
+import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.Gallery;
+import com.codestates.mainproject.oneyearfourcut.domain.member.entity.Member;
+import com.codestates.mainproject.oneyearfourcut.domain.member.repository.MemberRepository;
+import com.codestates.mainproject.oneyearfourcut.global.config.auth.jwt.PrincipalDto;
+import com.codestates.mainproject.oneyearfourcut.global.page.*;
+import com.google.gson.Gson;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
+import org.springframework.http.MediaType;
+import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
+import org.springframework.restdocs.payload.JsonFieldType;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultActions;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import static com.codestates.mainproject.oneyearfourcut.global.util.ApiDocumentUtils.getRequestPreProcessor;
+import static com.codestates.mainproject.oneyearfourcut.global.util.ApiDocumentUtils.getResponsePreProcessor;
+import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willDoNothing;
+import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebMvcTest(CommentController.class)
+@MockBean({JpaMetamodelMappingContext.class, ClientRegistrationRepository.class, MemberRepository.class})
+@AutoConfigureRestDocs
+public class CommentControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockBean
+    private CommentService commentService;
+
+    @Autowired
+    private Gson gson;
+
+    LocalDateTime time = LocalDateTime.of(2022, 12, 25, 12, 25, 25);
+
+    @TestConfiguration
+    static class testSecurityConfig {
+
+        @Bean
+        public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+            return http
+                    .csrf().disable()
+                    .build();
+        }
+    }
+
+    @BeforeEach
+    public void setup() {
+        //security context holder
+        String username = "test";
+        long id = 1L;
+        List<GrantedAuthority> authorities = List.of(new SimpleGrantedAuthority("ROLE_USER"));
+        Authentication authentication = new UsernamePasswordAuthenticationToken(new PrincipalDto(username, id), null, authorities);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+
+    @Test
+    void postCommentOnGalleryTest() throws Exception {
+        Gallery gallery = new Gallery(1L);
+
+        CommentRequestDto request = CommentRequestDto
+                .builder()
+                .content("텍스트")
+                .build();
+
+        CommentGalleryResDto responseDto = CommentGalleryResDto.builder()
+                .commentId(1L)
+                .createdAt(time)
+                .modifiedAt(time)
+                .memberId(1L)
+                .nickname("사용자")
+                .content(request.getContent())
+                .build();
+
+        CommentGalleryHeadDto<Object> response = new CommentGalleryHeadDto<>(gallery.getGalleryId(), responseDto);
+
+        given(commentService.createCommentOnGallery(any(CommentRequestDto.class), eq(gallery.getGalleryId()), anyLong()))
+                .willReturn(response);
+
+        String gsonContent = gson.toJson(request);
+
+        ResultActions actions = mockMvc.perform(
+                RestDocumentationRequestBuilders.post("/galleries/{galleryId}/comments", gallery.getGalleryId())
+                        .header("Authorization", "Bearer (AccessToken)")
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(gsonContent)
+        );
+
+        actions.andExpect(status().isCreated())
+                .andExpect(jsonPath("$.commentList.commentId").value(responseDto.getCommentId()))
+                .andExpect(jsonPath("$.commentList.content").value(responseDto.getContent()))
+                .andExpect(jsonPath("$.commentList.createdAt").value(String.valueOf(responseDto.getCreatedAt())))
+                .andExpect(jsonPath("$.commentList.modifiedAt").value(String.valueOf(responseDto.getModifiedAt())))
+                .andExpect(jsonPath("$.commentList.memberId").value(responseDto.getMemberId()))
+                .andExpect(jsonPath("$.commentList.nickname").value(responseDto.getNickname()))
+                .andExpect(jsonPath("$.galleryId").value(gallery.getGalleryId()))
+                .andExpect(jsonPath("$.commentList.artworkId").doesNotExist())
+                .andDo(document(
+                                "postCommentOnGallery",
+                                getRequestPreProcessor(),
+                                getResponsePreProcessor(),
+                                requestHeaders(
+                                        List.of(
+                                                headerWithName("Authorization").description("JWT - Access Token")
+                                        )
+                                ),
+                                pathParameters(
+                                        parameterWithName("galleryId").description("갤러리 Id")),
+                                requestFields(
+                                        List.of(
+                                                fieldWithPath("content").type(JsonFieldType.STRING).description("댓글 내용")
+                                        )
+                                )
+                                , responseFields(
+                                        List.of(
+                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
+                                                fieldWithPath("commentList.createdAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList.modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList.commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
+                                                fieldWithPath("commentList.memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
+                                                fieldWithPath("commentList.nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
+                                                fieldWithPath("commentList.content").type(JsonFieldType.STRING).description("댓글 내용"),
+                                                fieldWithPath("commentList.artworkId").type(JsonFieldType.NUMBER).description("작품 ID").optional(),
+                                                fieldWithPath("commentList.imagePath").type(JsonFieldType.NULL).description("작품 이미지 url")
+                                        )
+                                )
+                        )
+                );
+    }
+
+    @Test
+    void postCommentOnArtworkTest() throws Exception {
+        Gallery gallery = new Gallery(1L);
+        Artwork artwork = new Artwork(1L);
+        CommentRequestDto request = CommentRequestDto.builder()
+                .content("텍스트")
+                .build();
+
+        CommentArtworkResDto responseDto = CommentArtworkResDto.builder()
+                .commentId(1L)
+                .createdAt(time)
+                .modifiedAt(time)
+                .memberId(1L)
+                .nickname("사용자")
+                .content("텍스트")
+                .build();
+        CommentArtworkHeadDto<Object> response = new CommentArtworkHeadDto<>(gallery.getGalleryId(), artwork.getArtworkId(), responseDto);
+
+        given(commentService.createCommentOnArtwork(any(CommentRequestDto.class), eq(gallery.getGalleryId()), anyLong(), eq(artwork.getArtworkId())))
+                .willReturn(response);
+
+        String gsonContent = gson.toJson(request);
+
+        //when
+        ResultActions actions = mockMvc.perform(
+                RestDocumentationRequestBuilders
+                        .post("/galleries/{galleryId}/artworks/{artworkId}/comments",
+                                gallery.getGalleryId(), artwork.getArtworkId())
+                        .header("Authorization", "Bearer (AccessToken)")
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(gsonContent)
+        );
+
+        actions.andExpect(status().isCreated())
+                .andExpect(jsonPath("$.commentList.commentId").value(responseDto.getCommentId()))
+                .andExpect(jsonPath("$.commentList.content").value(responseDto.getContent()))
+                .andExpect(jsonPath("$.commentList.createdAt").value(String.valueOf(responseDto.getCreatedAt())))
+                .andExpect(jsonPath("$.commentList.modifiedAt").value(String.valueOf(responseDto.getModifiedAt())))
+                .andExpect(jsonPath("$.commentList.memberId").value(responseDto.getMemberId()))
+                .andExpect(jsonPath("$.commentList.nickname").value(responseDto.getNickname()))
+                .andExpect(jsonPath("$.galleryId").value(gallery.getGalleryId()))
+                .andExpect(jsonPath("$.artworkId").value(artwork.getArtworkId()))
+
+                .andDo(document(
+                                "postCommentOnArtwork",
+                                getRequestPreProcessor(),
+                                getResponsePreProcessor(),
+                                requestHeaders(
+                                        List.of(
+                                                headerWithName("Authorization").description("JWT - Access Token")
+                                        )
+                                )
+                                , pathParameters(
+                                        parameterWithName("galleryId").description("갤러리 Id"),
+                                        parameterWithName("artworkId").description("작품 Id")),
+
+                                requestFields(
+                                        List.of(
+                                                fieldWithPath("content").type(JsonFieldType.STRING).description("댓글 내용")
+                                        )
+                                )
+                                , responseFields(
+                                        List.of(
+                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
+                                                fieldWithPath("artworkId").type(JsonFieldType.NUMBER).description("작품 ID"),
+                                                fieldWithPath("commentList.createdAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList.modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList.commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
+                                                fieldWithPath("commentList.memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
+                                                fieldWithPath("commentList.nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
+                                                fieldWithPath("commentList.content").type(JsonFieldType.STRING).description("댓글 내용")
+                                        )
+                                )
+                        )
+                );
+
+    }
+
+    @Test
+    void getGalleryCommentTest() throws Exception {
+        Gallery gallery = new Gallery(1L);
+        int page = 1;
+        int size = 10;
+        PageInfo<Object> pageInfo = new PageInfo<>(page, size, 4, 1);
+        List<CommentGalleryResDto> collect = List.of(
+                CommentGalleryResDto.builder()
+                        .commentId(1L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(1L)
+                        .nickname("사용자1")
+                        .content("댓글1")
+                        .build(),
+                CommentGalleryResDto.builder()
+                        .commentId(2L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(2L)
+                        .nickname("사용자2")
+                        .content("댓글2")
+                        .artworkId(1L)
+                        .imagePath("artwork1.jpg").build(),
+                CommentGalleryResDto.builder()
+                        .commentId(3L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(3L)
+                        .nickname("사용자3")
+                        .content("댓글3")
+                        .build(),
+                CommentGalleryResDto.builder()
+                        .commentId(4L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(4L)
+                        .nickname("사용자4")
+                        .content("댓글4")
+                        .build());
+        CommentGalleryPageResponseDto<Object> response = new CommentGalleryPageResponseDto<>(gallery.getGalleryId(), collect, pageInfo);
+        given(commentService.getGalleryCommentPage(eq(gallery.getGalleryId()), eq(page), eq(size))).willReturn(response);
+
+        ResultActions actions = mockMvc.perform(
+                RestDocumentationRequestBuilders
+                        .get("/galleries/{galleryId}/comments",
+                                gallery.getGalleryId())
+                        .param("page", String.valueOf(1))
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+        );
+        commentService.getGalleryCommentPage(gallery.getGalleryId(), 1, 10);
+
+        //then
+        actions.andExpect(status().isOk())
+                .andExpect(jsonPath("$.galleryId").value(gallery.getGalleryId()))
+                .andExpect(jsonPath("$.commentList[0].createdAt").value(String.valueOf(collect.get(0).getCreatedAt())))
+                .andExpect(jsonPath("$.commentList[0].modifiedAt").value(String.valueOf(collect.get(0).getModifiedAt())))
+                .andExpect(jsonPath("$.commentList[0].commentId").value(collect.get(0).getCommentId()))
+                .andExpect(jsonPath("$.commentList[0].memberId").value(collect.get(0).getMemberId()))
+                .andExpect(jsonPath("$.commentList[0].nickname").value(collect.get(0).getNickname()))
+                .andExpect(jsonPath("$.commentList[0].content").value(collect.get(0).getContent()))
+                .andExpect(jsonPath("$.commentList[0].artworkId").doesNotExist())
+                .andExpect(jsonPath("$.commentList[0].imagePath").doesNotExist())
+
+                .andExpect(jsonPath("$.commentList[1].createdAt").value(String.valueOf(collect.get(1).getCreatedAt())))
+                .andExpect(jsonPath("$.commentList[1].modifiedAt").value(String.valueOf(collect.get(1).getModifiedAt())))
+                .andExpect(jsonPath("$.commentList[1].commentId").value(collect.get(1).getCommentId()))
+                .andExpect(jsonPath("$.commentList[1].memberId").value(collect.get(1).getMemberId()))
+                .andExpect(jsonPath("$.commentList[1].nickname").value(collect.get(1).getNickname()))
+                .andExpect(jsonPath("$.commentList[1].content").value(collect.get(1).getContent()))
+                .andExpect(jsonPath("$.commentList[1].artworkId").value(collect.get(1).getArtworkId()))
+                .andExpect(jsonPath("$.commentList[1].imagePath").value(collect.get(1).getImagePath()))
+
+                .andExpect(jsonPath("$.pageInfo.page").value(pageInfo.getPage()))
+                .andExpect(jsonPath("$.pageInfo.size").value(pageInfo.getSize()))
+                .andExpect(jsonPath("$.pageInfo.totalElements").value(pageInfo.getTotalElements()))
+                .andExpect(jsonPath("$.pageInfo.totalPages").value(pageInfo.getTotalPages()))
+
+                .andDo(document(
+                                "getCommentOnGallery",
+                                getRequestPreProcessor(),
+                                getResponsePreProcessor(),
+                                requestParameters(parameterWithName("page").description("페이지 수")
+                                )
+                                , pathParameters(
+                                        parameterWithName("galleryId").description("갤러리 Id"))
+                                , responseFields(
+                                        List.of(
+                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
+                                                fieldWithPath("commentList[].createdAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList[].modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList[].commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
+                                                fieldWithPath("commentList[].memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
+                                                fieldWithPath("commentList[].nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
+                                                fieldWithPath("commentList[].content").type(JsonFieldType.STRING).description("댓글 내용"),
+                                                fieldWithPath("commentList[].artworkId").type(JsonFieldType.NUMBER).description("작품 ID").optional(),
+                                                fieldWithPath("commentList[].imagePath").type(JsonFieldType.STRING).description("작품 이미지 URL (없으면 null)").optional(),
+
+                                                fieldWithPath("pageInfo.page").type(JsonFieldType.NUMBER).description("현 페이지 위치"),
+                                                fieldWithPath("pageInfo.size").type(JsonFieldType.NUMBER).description("페이지 내용 크기"),
+                                                fieldWithPath("pageInfo.totalElements").type(JsonFieldType.NUMBER).description("총 댓글 수"),
+                                                fieldWithPath("pageInfo.totalPages").type(JsonFieldType.NUMBER).description("총 페이지 수")
+                                        )
+                                )
+                        )
+                );
+    }
+
+    @Test
+    void getArtworkCommentTest() throws Exception {
+        Gallery gallery = new Gallery(1L);
+        Artwork artwork = new Artwork(1L);
+        int page = 1;
+        int size = 10;
+        PageInfo<Object> pageInfo = new PageInfo<>(page, size, 4, 1);
+        List<CommentArtworkResDto> collect = List.of(
+                CommentArtworkResDto.builder()
+                        .commentId(1L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(1L)
+                        .nickname("사용자1")
+                        .content("댓글1")
+                        .build(),
+                CommentArtworkResDto.builder()
+                        .commentId(2L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(2L)
+                        .nickname("사용자2")
+                        .content("댓글2")
+                        .build(),
+                CommentArtworkResDto.builder()
+                        .commentId(3L)
+                        .createdAt(time)
+                        .modifiedAt(time)
+                        .memberId(3L)
+                        .nickname("사용자3")
+                        .content("댓글3")
+                        .build(),
+                CommentArtworkResDto.builder()
+                        .commentId(4L)
+                        .createdAt(LocalDateTime.now())
+                        .modifiedAt(LocalDateTime.now())
+                        .memberId(4L)
+                        .nickname("사용자4")
+                        .content("댓글4")
+                        .build());
+        CommentArtworkPageResponseDto<Object> response = new CommentArtworkPageResponseDto<>(gallery.getGalleryId(), artwork.getArtworkId(), collect, pageInfo);
+        given(commentService.getArtworkCommentPage(eq(gallery.getGalleryId()), eq(artwork.getArtworkId()), eq(page), eq(size))).willReturn(response);
+
+        ResultActions actions = mockMvc.perform(
+                RestDocumentationRequestBuilders
+                        .get("/galleries/{galleryId}/artworks/{artworkId}/comments",
+                                gallery.getGalleryId(), artwork.getArtworkId())
+                        .param("page", String.valueOf(1))
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+                /*.with(csrf())*/
+        );
+        commentService.getArtworkCommentPage(gallery.getGalleryId(), 1L, 1, 10);
+
+        //then
+        actions.andExpect(status().isOk())
+                .andExpect(jsonPath("$.galleryId").value(gallery.getGalleryId()))
+                .andExpect(jsonPath("$.artworkId").value(artwork.getArtworkId()))
+                .andExpect(jsonPath("$.commentList[0].commentId").value(collect.get(0).getCommentId()))
+                .andExpect(jsonPath("$.commentList[0].memberId").value(collect.get(0).getMemberId()))
+                .andExpect(jsonPath("$.commentList[0].nickname").value(collect.get(0).getNickname()))
+                .andExpect(jsonPath("$.commentList[0].content").value(collect.get(0).getContent()))
+
+                .andExpect(jsonPath("$.pageInfo.page").value(pageInfo.getPage()))
+                .andExpect(jsonPath("$.pageInfo.size").value(pageInfo.getSize()))
+                .andExpect(jsonPath("$.pageInfo.totalElements").value(pageInfo.getTotalElements()))
+                .andExpect(jsonPath("$.pageInfo.totalPages").value(pageInfo.getTotalPages()))
+
+
+                .andDo(document(
+                                "getCommentOnArtwork",
+                                getRequestPreProcessor(),
+                                getResponsePreProcessor(),
+                                requestParameters(parameterWithName("page").description("페이지 수")
+                                )
+                                , pathParameters(
+                                        parameterWithName("galleryId").description("갤러리 Id"),
+                                        parameterWithName("artworkId").description("작품 Id"))
+                                , responseFields(
+                                        List.of(
+                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
+                                                fieldWithPath("artworkId").type(JsonFieldType.NUMBER).description("작품 ID"),
+                                                fieldWithPath("commentList[].createdAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList[].modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList[].commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
+                                                fieldWithPath("commentList[].memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
+                                                fieldWithPath("commentList[].nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
+                                                fieldWithPath("commentList[].content").type(JsonFieldType.STRING).description("댓글 내용"),
+
+                                                fieldWithPath("pageInfo.page").type(JsonFieldType.NUMBER).description("현 페이지 위치"),
+                                                fieldWithPath("pageInfo.size").type(JsonFieldType.NUMBER).description("페이지 내용 크기"),
+                                                fieldWithPath("pageInfo.totalElements").type(JsonFieldType.NUMBER).description("총 댓글 수"),
+                                                fieldWithPath("pageInfo.totalPages").type(JsonFieldType.NUMBER).description("총 페이지 수")
+                                        )
+                                )
+                        )
+                );
+    }
+
+    @Test
+    void patchCommentTest() throws Exception {
+        Gallery gallery = new Gallery(1L);
+        CommentRequestDto requestDto = CommentRequestDto.builder().content("수정된 댓글").build();
+
+        CommentGalleryResDto responseDto = CommentGalleryResDto.builder()
+                .commentId(1L)
+                .createdAt(time)
+                .modifiedAt(time)
+                .memberId(1L)
+                .nickname("사용자")
+                .content(requestDto.getContent())
+                .build();
+
+        CommentGalleryHeadDto<Object> response = new CommentGalleryHeadDto<>(gallery.getGalleryId(), responseDto);
+        given(commentService.modifyComment(eq(gallery.getGalleryId()), eq(responseDto.getCommentId()), any(CommentRequestDto.class), anyLong()))
+                .willReturn(response);
+
+        String gsonContent = gson.toJson(requestDto);
+
+        ResultActions actions = mockMvc.perform(
+                RestDocumentationRequestBuilders
+                        .patch("/galleries/{galleryId}/comments/{commentId}",
+                                gallery.getGalleryId(), responseDto.getCommentId())
+                        .header("Authorization", "Bearer (AccessToken)")
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(gsonContent)
+        );
+        actions.andExpect(status().isOk())
+                .andExpect(jsonPath("$.galleryId").value(gallery.getGalleryId()))
+                .andExpect(jsonPath("$.commentList.commentId").value(responseDto.getCommentId()))
+                .andExpect(jsonPath("$.commentList.memberId").value(responseDto.getMemberId()))
+                .andExpect(jsonPath("$.commentList.nickname").value(responseDto.getNickname()))
+                .andExpect(jsonPath("$.commentList.content").value(responseDto.getContent()))
+                .andExpect(jsonPath("$.commentList.artworkId").doesNotExist())
+
+                .andDo(document(
+                                "patchComment",
+                                getRequestPreProcessor(),
+                                getResponsePreProcessor(),
+                                requestHeaders(
+                                        List.of(
+                                                headerWithName("Authorization").description("JWT - Access Token")
+                                        )
+                                ),
+                                pathParameters(
+                                        parameterWithName("galleryId").description("갤러리 Id"),
+                                        parameterWithName("commentId").description("댓글 Id")),
+                                requestFields(
+                                        List.of(
+                                                fieldWithPath("content").type(JsonFieldType.STRING).description("댓글 내용")
+                                        )
+                                )
+                                , responseFields(
+                                        List.of(
+                                                fieldWithPath("galleryId").type(JsonFieldType.NUMBER).description("전시관 ID"),
+                                                fieldWithPath("commentList.createdAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList.modifiedAt").type(JsonFieldType.STRING).description("생성일자"),
+                                                fieldWithPath("commentList.commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
+                                                fieldWithPath("commentList.memberId").type(JsonFieldType.NUMBER).description("회원 ID"),
+                                                fieldWithPath("commentList.nickname").type(JsonFieldType.STRING).description("회원 닉네임"),
+                                                fieldWithPath("commentList.content").type(JsonFieldType.STRING).description("댓글 내용"),
+                                                fieldWithPath("commentList.artworkId").type(JsonFieldType.NUMBER).description("작품 ID").optional(),
+                                                fieldWithPath("commentList.imagePath").type(JsonFieldType.NULL).description("작품 이미지 url").optional()
+                                        )
+                                )
+                        )
+                );
+    }
+
+    @Test
+    void deleteCommentTest() throws Exception {
+        Member member = new Member(1L);
+        Gallery gallery = new Gallery(1L);
+        Comment comment = new Comment(1L);
+
+        willDoNothing().given(commentService).deleteComment(eq(gallery.getGalleryId()), eq(comment.getCommentId()), eq(member.getMemberId()));
+
+        ResultActions actions = mockMvc.perform(
+                RestDocumentationRequestBuilders
+                        .delete("/galleries/{galleryId}/comments/{commentId}",
+                                gallery.getGalleryId(), comment.getCommentId())
+                        .header("Authorization", "Bearer (AccessToken)")
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+        );
+        actions.andExpect(status().isNoContent())
+                .andDo(document(
+                                "deleteComment",
+                                getRequestPreProcessor(),
+                                getResponsePreProcessor(),
+                                requestHeaders(
+                                        List.of(
+                                                headerWithName("Authorization").description("JWT - Access Token")
+                                        )
+                                ),
+                                pathParameters(
+                                        parameterWithName("galleryId").description("갤러리 Id"),
+                                        parameterWithName("commentId").description("댓글 Id"))
+                        )
+
+                );
+
+
+    }
+}
