+++ server/oneyearfourcut/src/test/java/com/codestates/mainproject/oneyearfourcut/domain/comment/controller/ReplyControllerTest.java
@@ -2,47 +2,41 @@ package com.codestates.mainproject.oneyearfourcut.domain.comment.controller;
 
 import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.CommentRequestDto;
 import com.codestates.mainproject.oneyearfourcut.domain.comment.dto.ReplyResDto;
-import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.Comment;
-import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.CommentStatus;
-import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.Reply;
-import com.codestates.mainproject.oneyearfourcut.domain.comment.repository.CommentRepository;
-import com.codestates.mainproject.oneyearfourcut.domain.comment.repository.ReplyRepository;
-import com.codestates.mainproject.oneyearfourcut.domain.comment.service.CommentService;
 import com.codestates.mainproject.oneyearfourcut.domain.comment.service.ReplyService;
-import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.Gallery;
-import com.codestates.mainproject.oneyearfourcut.domain.gallery.entity.GalleryStatus;
-import com.codestates.mainproject.oneyearfourcut.domain.gallery.repository.GalleryRepository;
-import com.codestates.mainproject.oneyearfourcut.domain.member.entity.Member;
-import com.codestates.mainproject.oneyearfourcut.domain.member.entity.Role;
 import com.codestates.mainproject.oneyearfourcut.domain.member.repository.MemberRepository;
-import com.codestates.mainproject.oneyearfourcut.global.config.auth.jwt.JwtTokenizer;
+import com.codestates.mainproject.oneyearfourcut.global.config.auth.jwt.PrincipalDto;
 import com.codestates.mainproject.oneyearfourcut.global.page.ReplyListResponseDto;
 import com.google.gson.Gson;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
-import static com.codestates.mainproject.oneyearfourcut.domain.comment.entity.CommentStatus.VALID;
-import static com.codestates.mainproject.oneyearfourcut.domain.member.entity.MemberStatus.ACTIVE;
 import static com.codestates.mainproject.oneyearfourcut.global.util.ApiDocumentUtils.getRequestPreProcessor;
 import static com.codestates.mainproject.oneyearfourcut.global.util.ApiDocumentUtils.getResponsePreProcessor;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willDoNothing;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
@@ -54,96 +48,79 @@ import static org.springframework.restdocs.request.RequestDocumentation.pathPara
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@SpringBootTest
-@Transactional
-@AutoConfigureMockMvc
+@WebMvcTest(ReplyController.class)
+@MockBean({JpaMetamodelMappingContext.class, ClientRegistrationRepository.class, MemberRepository.class})
 @AutoConfigureRestDocs
-@WithMockUser(username = "test@gmail.com", password = "0000")
-class ReplyControllerTest {
+public class ReplyControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
-    @Autowired
-    private JwtTokenizer jwtTokenizer;
-    @Autowired
-    private MemberRepository memberRepository;
-    @Autowired
-    private GalleryRepository galleryRepository;
-    @Autowired
-    private CommentRepository commentRepository;
-    @Autowired
-    private ReplyRepository replyRepository;
+
     @MockBean
     private ReplyService replyService;
+
     @Autowired
     private Gson gson;
 
+    LocalDateTime time = LocalDateTime.of(2022, 12, 25, 12, 25, 25);
+
+    @TestConfiguration
+    static class testSecurityConfig {
+
+        @Bean
+        public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+            return http
+                    .csrf().disable()
+                    .build();
+        }
+    }
+
+    @BeforeEach
+    public void setup() {
+        //security context holder
+        String username = "test";
+        long id = 1L;
+        List<GrantedAuthority> authorities = List.of(new SimpleGrantedAuthority("ROLE_USER"));
+        Authentication authentication = new UsernamePasswordAuthenticationToken(new PrincipalDto(username, id), null, authorities);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
 
     @Test
-    void testPostReply() throws Exception {
-        //given
-        Member member = memberRepository.save(Member.builder()
-                .nickname("test1")
-                .email("kang@gmail.com")
-                .role(Role.USER)
-                .profile("/path")
-                .status(ACTIVE)
-                .build());
-        String jwt = jwtTokenizer.testJwtGenerator(member);
-        Gallery gallery = galleryRepository.save(new Gallery(1L));
+    void postReplyTest() throws Exception {
+        Long commentId = 1L;
+        Long memberId = 1L;
 
-        CommentRequestDto requestDto = CommentRequestDto.builder()
-                .content("답글입니다")
-                .build();
-        String gsonContent = gson.toJson(requestDto);
+        CommentRequestDto requestDto = CommentRequestDto.builder().content("대댓글").build();
 
         ReplyResDto responseDto = ReplyResDto.builder()
-                .createdAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-                .modifiedAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
                 .replyId(1L)
-                .memberId(member.getMemberId())
-                .nickname(member.getNickname())
+                .memberId(memberId)
+                .nickname("작성자")
                 .content(requestDto.getContent())
-                .build();
+                .createdAt(time)
+                .modifiedAt(time).build();
+        ReplyListResponseDto<Object> response = new ReplyListResponseDto<>(commentId, responseDto);
+        System.out.println(response.getCommentId());
+        given(replyService.createReply(any(CommentRequestDto.class), eq(commentId), eq(memberId))).willReturn(response);
 
-        Comment comment = commentRepository.save(Comment.builder()
-                .commentId(1L)
-                .content("댓글")
-                .member(member)
-                .gallery(gallery)
-                .artworkId(1L)
-                .commentStatus(CommentStatus.VALID)
-                .build());
-
-
-        given(this.replyService.createReply(
-                any(requestDto.getClass()),
-                any( comment.getCommentId().getClass()),
-                any( member.getMemberId().getClass() )))
-                .willReturn(new ReplyListResponseDto<>(1L,responseDto));
-
-        //when
+        String gsonContent = gson.toJson(requestDto);
         ResultActions actions = mockMvc.perform(
                 RestDocumentationRequestBuilders
                         .post("/galleries/comments/{commentId}/replies",
-                                comment.getCommentId())
-                        .header("Authorization", jwt)
+                                commentId)
+                        .header("Authorization", "Bearer (AccessToken)")
                         .accept(MediaType.APPLICATION_JSON)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(gsonContent)
-                /*.with(csrf())*/
         );
-        replyService.createReply(requestDto, comment.getCommentId(),member.getMemberId());
-
-        //then
         actions.andExpect(status().isCreated())
-                .andExpect(jsonPath("$.commentId").value( comment.getCommentId()))
-                .andExpect(jsonPath("$.replyList.createdAt").value("2022-11-25T11:09:24.94"))
-                .andExpect(jsonPath("$.replyList.modifiedAt").value("2022-11-25T11:09:24.94"))
-                .andExpect(jsonPath("$.replyList.replyId").value(1L))
-                .andExpect(jsonPath("$.replyList.memberId").value(member.getMemberId()))
-                .andExpect(jsonPath("$.replyList.nickname").value("test1"))
-                .andExpect(jsonPath("$.replyList.content").value("답글입니다"))
+                .andExpect(jsonPath("$.commentId").value(response.getCommentId()))
+                .andExpect(jsonPath("$.replyList.createdAt").value(responseDto.getCreatedAt().toString()))
+                .andExpect(jsonPath("$.replyList.modifiedAt").value(responseDto.getModifiedAt().toString()))
+                .andExpect(jsonPath("$.replyList.replyId").value(responseDto.getReplyId()))
+                .andExpect(jsonPath("$.replyList.memberId").value(responseDto.getMemberId()))
+                .andExpect(jsonPath("$.replyList.nickname").value(responseDto.getNickname()))
+                .andExpect(jsonPath("$.replyList.content").value(responseDto.getContent()))
 
                 .andDo(document(
                                 "postReply",
@@ -159,8 +136,8 @@ class ReplyControllerTest {
                                                 fieldWithPath("content").type(JsonFieldType.STRING).description("답글 내용")
                                         )
                                 )
-                        ,pathParameters(
-                                parameterWithName("commentId").description("댓글 Id"))
+                                ,pathParameters(
+                                        parameterWithName("commentId").description("댓글 Id"))
                                 , responseFields(
                                         List.of(
                                                 fieldWithPath("commentId").type(JsonFieldType.NUMBER).description("댓글 ID"),
@@ -175,67 +152,53 @@ class ReplyControllerTest {
                                 )
                         )
                 );
-
     }
+
     @Test
-    void testGetReply() throws Exception {
-        //given
-        Member member = memberRepository.save(Member.builder()
-                .nickname("kang")
-                .email("kang@gmail.com")
-                .role(Role.USER)
-                .profile("/path")
-                .status(ACTIVE)
-                .build());
-        Gallery gallery = galleryRepository.save(new Gallery(1L));
+    void getReplyTest() throws Exception {
+        Long commentId = 1L;
 
-        Comment comment = commentRepository.save(Comment.builder()
-                .commentId(1L)
-                .content("댓글")
-                .member(member)
-                .gallery(gallery)
-                .artworkId(1L)
-                .commentStatus(CommentStatus.VALID)
-                .build());
+        List<ReplyResDto> responseDtoList = List.of(
+                ReplyResDto.builder()
+                        .replyId(1L)
+                        .memberId(1L)
+                        .nickname("작성자1")
+                        .content("대댓글1")
+                        .createdAt(time)
+                        .modifiedAt(time).build(),
+                ReplyResDto.builder()
+                        .replyId(2L)
+                        .memberId(2L)
+                        .nickname("작성자2")
+                        .content("대댓글2")
+                        .createdAt(time)
+                        .modifiedAt(time).build(),
+                ReplyResDto.builder()
+                        .replyId(3L)
+                        .memberId(3L)
+                        .nickname("작성자3")
+                        .content("대댓글3")
+                        .createdAt(time)
+                        .modifiedAt(time).build());
 
-        Reply reply = replyRepository.save(Reply.builder()
-                .replyId(4L)
-                .content("답글")
-                .member(member)
-                .comment(comment)
-                .replyStatus(CommentStatus.VALID)
-                .build());
+        ReplyListResponseDto<Object> response = new ReplyListResponseDto<>(commentId, responseDtoList);
 
-        List<Reply> replyList1 = new ArrayList<>();
-        replyList1.add(reply);
-        given(replyService.findReplyList(Mockito.any( comment.getCommentId().getClass())))
-                .willReturn(replyList1);
+        given(replyService.getReplyList(eq(commentId))).willReturn(response);
 
-        List<Reply> replyList = replyService.findReplyList(1L);
-        List<ReplyResDto> result = ReplyResDto.toReplyResponseDtoList(replyList);
-
-        given(this.replyService.getReplyList(
-                any( comment.getCommentId().getClass())))
-                .willReturn(new ReplyListResponseDto<>(1L,result));
-
-        //when
         ResultActions actions = mockMvc.perform(
                 RestDocumentationRequestBuilders
                         .get("/galleries/comments/{commentId}/replies",
-                                comment.getCommentId())
+                                commentId)
                         .accept(MediaType.APPLICATION_JSON)
                         .contentType(MediaType.APPLICATION_JSON)
-                        /*.with(csrf())*/
         );
-        replyService.getReplyList(comment.getCommentId());
 
-        //then
         actions.andExpect(status().isOk())
-                .andExpect(jsonPath("$.commentId").value( comment.getCommentId()))
-                .andExpect(jsonPath("$.replyList[0].replyId").value(4L))
-                .andExpect(jsonPath("$.replyList[0].memberId").value(member.getMemberId()))
-                .andExpect(jsonPath("$.replyList[0].nickname").value("kang"))
-                .andExpect(jsonPath("$.replyList[0].content").value("답글"))
+                .andExpect(jsonPath("$.commentId").value(response.getCommentId()))
+                .andExpect(jsonPath("$.replyList[0].replyId").value(responseDtoList.get(0).getReplyId()))
+                .andExpect(jsonPath("$.replyList[0].memberId").value(responseDtoList.get(0).getMemberId()))
+                .andExpect(jsonPath("$.replyList[0].nickname").value(responseDtoList.get(0).getNickname()))
+                .andExpect(jsonPath("$.replyList[0].content").value(responseDtoList.get(0).getContent()))
 
                 .andDo(document(
                                 "getReply",
@@ -259,100 +222,42 @@ class ReplyControllerTest {
                 );
 
     }
+
     @Test
-    void testPatchReply() throws Exception{
-        //given
-        Member member = memberRepository.save(Member.builder()
-                .nickname("test1")
-                .email("kang@gmail.com")
-                .role(Role.USER)
-                .profile("/path")
-                .status(ACTIVE)
-                .build());
-        String jwt = jwtTokenizer.testJwtGenerator(member);
-
-        Gallery gallery = galleryRepository.save(Gallery.builder()
-                .title("나의 전시관")
-                .content("안녕하세요")
-                .member(member)
-                .status(GalleryStatus.OPEN)
-                .build());
-
-        Comment comment = commentRepository.save(Comment.builder()
-                .commentId(4L)
-                .content("댓글")
-                .member(member)
-                .gallery(gallery)
-                .artworkId(null)
-                .commentStatus(CommentStatus.VALID)
-                .build());
-
-        replyRepository.save(Reply.builder()
-                .replyId(1L)
-                .content("답글1")
-                .member(member)
-                .comment(comment)
-                .replyStatus(CommentStatus.VALID)
-                .build());
-
-        replyRepository.save(Reply.builder()
-                .replyId(2L)
-                .content("답글2")
-                .member(member)
-                .comment(comment)
-                .replyStatus(CommentStatus.VALID)
-                .build());
-
-        CommentRequestDto requestDto = CommentRequestDto.builder()
-                .content("수정대댓글입니다.")
-                .build();
-        String gsonContent = gson.toJson(requestDto);
-
+    void patchReplyTest() throws Exception {
+        Long commentId = 1L;
+        Long memberId = 1L;
         Long replyId = 1L;
-
-        given(this.replyService.findReplyList(
-                any( comment.getCommentId().getClass() )))
-                .willReturn(replyRepository.findAllByReplyStatusAndComment_CommentIdOrderByReplyIdDesc(VALID, 3L));
-
-        List<Reply> replyList = replyService.findReplyList(3L);
-        List<ReplyResDto> result = ReplyResDto.toReplyResponseDtoList(replyList);
+        CommentRequestDto requestDto = CommentRequestDto.builder().content("수정대댓글").build();
 
         ReplyResDto responseDto = ReplyResDto.builder()
-                .createdAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-                .modifiedAt(LocalDateTime.parse("2022-11-25T11:09:24.940"))
-                .replyId(replyId)
-                .memberId(member.getMemberId())
-                .nickname(member.getNickname())
+                .replyId(1L)
+                .memberId(memberId)
+                .nickname("작성자")
                 .content(requestDto.getContent())
-                .build();
+                .createdAt(time)
+                .modifiedAt(time).build();
+        ReplyListResponseDto<Object> response = new ReplyListResponseDto<>(commentId, responseDto);
 
-        given(this.replyService.modifyReply(
-                any( comment.getCommentId().getClass()),
-                eq(replyId),
-                any( requestDto.getClass() ),
-                any( member.getMemberId().getClass() )))
-                .willReturn(new ReplyListResponseDto<>( gallery.getGalleryId() , responseDto));
+        given(replyService.modifyReply(eq(commentId), eq(replyId), any(CommentRequestDto.class), eq(memberId))).willReturn(response);
 
-        replyService.modifyReply(4L,1L, requestDto ,member.getMemberId());
+        String gsonContent = gson.toJson(requestDto);
 
-        //when
         ResultActions actions = mockMvc.perform(
                 RestDocumentationRequestBuilders
                         .patch("/galleries/comments/{commentId}/replies/{replyId}",
-                               4L, 1L)
-                        .header("Authorization", jwt)
+                                commentId, replyId)
+                        .header("Authorization", "Bearer (AccessToken)")
                         .accept(MediaType.APPLICATION_JSON)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(gsonContent)
-                /*.with(csrf())*/
         );
 
-        //then
         actions.andExpect(status().isOk())
-                .andExpect(jsonPath("$.commentId").value( comment.getCommentId()))
-                .andExpect(jsonPath("$.replyList.memberId").value(member.getMemberId()))
-                .andExpect(jsonPath("$.replyList.nickname").value("test1"))
-                .andExpect(jsonPath("$.replyList.content").value("수정대댓글입니다."))
+                .andExpect(jsonPath("$.commentId").value(response.getCommentId()))
+                .andExpect(jsonPath("$.replyList.memberId").value(responseDto.getMemberId()))
+                .andExpect(jsonPath("$.replyList.nickname").value(responseDto.getNickname()))
+                .andExpect(jsonPath("$.replyList.content").value(responseDto.getContent()))
 
                 .andDo(document(
                                 "patchReply",
@@ -385,48 +290,20 @@ class ReplyControllerTest {
                         )
                 );
     }
+
     @Test
-    void testDeleteReply() throws Exception{
-        //given
-        Member member = memberRepository.save(Member.builder()
-                .nickname("test1")
-                .email("kang@gmail.com")
-                .role(Role.USER)
-                .profile("/path")
-                .status(ACTIVE)
-                .build());
-        String jwt = jwtTokenizer.testJwtGenerator(member);
+    void deleteReplyTest() throws Exception {
+        Long commentId = 1L;
+        Long memberId = 1L;
+        Long replyId = 1L;
 
-        Gallery gallery = galleryRepository.save(Gallery.builder()
-                .title("나의 전시관")
-                .content("안녕하세요")
-                .member(member)
-                .status(GalleryStatus.OPEN)
-                .build());
+        willDoNothing().given(replyService).deleteReply(eq(commentId), eq(replyId), eq(memberId));
 
-        Comment comment = commentRepository.save(Comment.builder()
-                .commentId(3L)
-                .content("댓글")
-                .member(member)
-                .gallery(gallery)
-                .artworkId(null)
-                .commentStatus(CommentStatus.VALID)
-                .build());
-
-        replyRepository.save(Reply.builder()
-                .replyId(3L)
-                .content("답글")
-                .member(member)
-                .comment(comment)
-                .replyStatus(CommentStatus.VALID)
-                .build());
-
-        //when
         ResultActions actions = mockMvc.perform(
                 RestDocumentationRequestBuilders
                         .delete("/galleries/comments/{commentId}/replies/{replyId}",
-                                3L, 3L)
-                        .header("Authorization", jwt)
+                                commentId, replyId)
+                        .header("Authorization", "Bearer (AccessToken)")
                         .accept(MediaType.APPLICATION_JSON)
                         .contentType(MediaType.APPLICATION_JSON)
                 /*.with(csrf())*/
@@ -447,9 +324,6 @@ class ReplyControllerTest {
                                         parameterWithName("commentId").description("댓글 Id"),
                                         parameterWithName("replyId").description("답글 Id"))
                         )
-
                 );
-
     }
-
 }
