+++ server/oneyearfourcut/src/test/java/com/codestates/mainproject/oneyearfourcut/domain/comment/repository/RepositoryTest.java
@@ -1,6 +1,7 @@
 package com.codestates.mainproject.oneyearfourcut.domain.comment.repository;
 
 
+import com.codestates.mainproject.oneyearfourcut.domain.artwork.entity.Artwork;
 import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.Comment;
 import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.CommentStatus;
 import com.codestates.mainproject.oneyearfourcut.domain.comment.entity.Reply;
@@ -55,8 +56,7 @@ class RepositoryTest {
                     .commentId(1L)
                     .member(new Member(3L))
                     .gallery(new Gallery(2L))
-                    .artworkId(133333L)
-                    .commentStatus(CommentStatus.DELETED)
+                    .artwork(new Artwork(1L))
                     .build()
         );
 
@@ -65,14 +65,17 @@ class RepositoryTest {
         List<Reply> replyList = replyRepository.findAll();
         Comment comment = commentRepository.findById(1L).orElse(null);
 
+
         //then
         Comment savedComment = commentList.get(0);
+        // 임시방편
+        savedComment.changeCommentStatus(CommentStatus.DELETED);
         Reply savedReply = replyList.get(0);
         assertThat(savedComment.getContent()).isEqualTo("this is sample test comment");
         assertThat(savedComment.getCommentId()).isEqualTo(1L);
         assertThat(savedComment.getMember().getMemberId()).isEqualTo(3L);
         assertThat(savedComment.getGallery().getGalleryId()).isEqualTo(2L);
-        assertThat(savedComment.getArtworkId()).isEqualTo(133333L);
+        assertThat(savedComment.getArtworkId()).isEqualTo(1L);
         assertThat(savedComment.getCommentStatus()).isEqualTo(CommentStatus.DELETED);
         assertThat(comment).isEqualTo(savedReply.getComment()); //mapping clear
